docs: Improve Readme (#30)
# Pull Request

## Description

This pull request updates the `README.md` to improve documentation and
make it more welcoming and informative for contributors. The changes add
sections for project inspiration, a more detailed contribution guide,
and clarify licensing.

Documentation improvements:

* Added a project description and a new "Inspiration" section
referencing similar projects to provide context for users and
contributors.
* Expanded the "Contributing" section with a quick step-by-step guide
and a list of ways to contribute, making it easier for newcomers to get
involved.
* Updated the Table of Contents and added emoji-enhanced headings for
better readability and navigation.
* Clarified the licensing section with a direct reference to the MIT
License file.