ci: Add CodeQL Go Job (#36)
# Pull Request

## Description

This pull request updates the code analysis workflow to include Go in
addition to Actions for CodeQL analysis. This change expands the scope
of static analysis to cover Go code in the repository.

Workflow improvements:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L62-R62):
Added `go` to the matrix of languages for CodeQL analysis, enabling
security and code quality checks for Go code.

Fixes #11