build: Add Commit Message and Output File Name Variables (#74)
# Pull Request

## Description

This pull request adds support for customizing the commit message when
updating the SVG output file. The main changes involve allowing users to
specify a commit message via workflow inputs and updating the code to
use this message during the commit process.

**Workflow input enhancements:**

* Added a new optional input `commit_message` to `action.yml`, allowing
users to specify a custom commit message for SVG updates.

**Code updates for commit customization:**

* Modified `src/commit.go` to read the `commit_message` input from the
environment and use it as the commit message instead of the previous
hardcoded value.
[[1]](diffhunk://#diff-d4596480fdccff713fe198cbf0bcf98788bf6179c05e3588799032ed6555921eL24-R25)
[[2]](diffhunk://#diff-d4596480fdccff713fe198cbf0bcf98788bf6179c05e3588799032ed6555921eL49-R50)
* Changed the file path assignment in `src/commit.go` to use the
`INPUT_OUTPUT_FILE_NAME` environment variable, ensuring consistency with
the workflow input for the output file name.