build: Get User Followers and Following Counts (#109)
# Pull Request

## Description

This pull request makes a minor update to how user information is
displayed in the SVG stats. The main change is that the number of users
a person is following is now dynamically shown based on their actual
GitHub data, instead of being hardcoded.

* Data model update: Added a new `Following` field to the
`GitHubUserInfo` struct to store the number of users a person is
following.
* Stats display improvement: Updated the SVG stats row to display the
actual number of users the person is following by referencing the new
`Following` field, instead of the previously hardcoded value.