docs: Improve documentation (#158)
# Pull Request

## Description

This pull request reorganizes contribution and testing documentation for
the project. The main change is moving the "Quick Contribution Guide"
from the `README.md` to a dedicated `docs/CONTRIBUTING.md` file, and
introducing a new "Testing" section in the `README.md` to clarify how
the GitHub Action is verified.

Documentation restructuring:

* The "Quick Contribution Guide" has been removed from the `README.md`
and added to `docs/CONTRIBUTING.md`, where it now includes a table of
contents for easier navigation.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L11-R11)
[[2]](diffhunk://#diff-4c6b93aa75d5affde60dc3849606c9acd75ed444d52e99f3055fc0c7aa77e9e0R4-R19)

Testing information:

* A new "Testing" section has been added to the `README.md`, describing
how the GitHub Action is regularly run in a test repository and
providing a link to the latest run result.