feat: Add yearly contribution graph (#182)
# Pull Request

## Description

This pull request updates the SVG creation logic to use constants for
the SVG canvas size, and increases the height of the SVG from 380 to 520
pixels. This change ensures consistency and makes it easier to adjust
the canvas size in the future.

SVG canvas size update:

* Introduced `svgWidth` and `svgHeight` constants, set to 1000 and 520
respectively, and replaced hardcoded dimensions in the background
rectangle and root SVG element with these constants. The SVG height was
increased from 380 to 520 pixels.